Finished tracking and changed everything to f64 from f32

+++ quadrupole_sim/src/main.rs
@@ -1,14 +1,14 @@
 mod physics;
 
 /// Constants in terms of Natural Units
-const PROTON_MASS: f32 = 938.7; // The mass of a proton in MeV/c^2
+const PROTON_MASS: f64 = 938.7; // The mass of a proton in MeV/c^2
 
 /// SI Units and Conversions
-const MU0: f32 = 1.256_637_061_4e-6; // Magnetic permeability T*m / A
-const C_TM: f32 = 299.792_458; // Speed of light c = 1 in Natural Units; conversion factor might be needed: MeV/T*m
+const MU0: f64 = 1.256_637_061_4e-6; // Magnetic permeability T*m / A
+const C_TM: f64 = 299.792_458; // Speed of light c = 1 in Natural Units; conversion factor might be needed: MeV/T*m
 
-const IN_TO_M: f32 = 0.0254; // Conversion factor 
-const MM_TO_M: f32 = 1e-3; // Conversion factor
+const IN_TO_M: f64 = 0.0254; // Conversion factor 
+const MM_TO_M: f64 = 1e-3; // Conversion factor
 
 fn main() -> std::io::Result<()> {
     Ok(())
